requirements.txt
Django==3.1.2
djangorestframework==3.12.4
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
requests==2.26.0
django-cors-headers==3.7.0